Migration to Hugo
Since Jekyll is for some reason not installable
anymore on my current version of OSX I decided to
try out something new, Hugo.

This is a migration that keeps the structure more
or less the same, but with a different layout.

The URLs to blog posts still work via aliases. I
don't think anyone has deep links to my blog, but
you never know. It's just good practice.

I use the cactus theme but forked it since some things
were not to my taste.

I also chose to keep the index page a bit shorter and
add an "about me" with more text.